+++ src/index/btree/mod.rs
@@ -40,9 +40,16 @@ use std::rc::Rc;
 use std::sync::atomic::{AtomicUsize, Ordering::Relaxed, Ordering::SeqCst};
 use std::sync::Arc;
 use utils::lru_cache::LRUCache;
+pub use index::btree::node::*;
+pub use index::btree::cursor::*;
+pub use index::btree::merge::*;
+
 
 mod external;
 mod internal;
+mod node;
+mod cursor;
+mod merge;
 
 pub const NUM_KEYS: usize = 4;
 const NUM_PTRS: usize = NUM_KEYS + 1;
@@ -66,58 +73,10 @@ pub struct BPlusTree {
 unsafe impl Sync for BPlusTree {}
 unsafe impl Send for BPlusTree {}
 
-// This is the runtime cursor on iteration
-// It hold a copy of the containing page next page lock guard
-// These lock guards are preventing the node and their neighbourhoods been changed externally
-// Ordering are specified that can also change lock pattern
-pub struct RTCursor {
-    index: usize,
-    ordering: Ordering,
-    page: Option<NodeCellRef>,
-}
-
 impl Default for Ordering {
     fn default() -> Self {
         Ordering::Forward
     }
-}
-
-enum InsertSearchResult {
-    External,
-    Internal(NodeCellRef),
-    RightNode(NodeCellRef),
-}
-
-enum InsertToNodeResult {
-    NoSplit,
-    Split(NodeWriteGuard, NodeCellRef, Option<EntryKey>),
-    SplitParentChanged,
-}
-
-enum RemoveSearchResult {
-    External,
-    RightNode(NodeCellRef),
-    Internal(NodeCellRef, SubNodeStatus),
-}
-
-enum SubNodeStatus {
-    ExtNodeEmpty,
-    InNodeEmpty,
-    Relocate(usize, usize),
-    Merge(usize, usize),
-    Ok,
-}
-
-pub struct NodeSplit {
-    new_right_node: NodeCellRef,
-    left_node_latch: NodeWriteGuard,
-    pivot: EntryKey,
-    parent_latch: Option<NodeWriteGuard>,
-}
-
-pub enum NodeSplitResult {
-    Split(NodeSplit),
-    Retry,
 }
 
 impl BPlusTree {
@@ -503,492 +462,6 @@ impl BPlusTree {
     }
 }
 
-pub enum NodeData {
-    External(Box<ExtNode>),
-    Internal(Box<InNode>),
-    None,
-}
-
-impl NodeData {
-    pub fn is_none(&self) -> bool {
-        match self {
-            &NodeData::None => true,
-            _ => false,
-        }
-    }
-    fn search(&self, key: &EntryKey) -> usize {
-        let len = self.len();
-        if self.is_ext() {
-            self.extnode().keys[..len]
-                .binary_search(key)
-                .unwrap_or_else(|i| i)
-        } else {
-            self.innode().keys[..len]
-                .binary_search(key)
-                .map(|i| i + 1)
-                .unwrap_or_else(|i| i)
-        }
-    }
-
-    fn remove(&mut self, pos: usize) {
-        match self {
-            &mut NodeData::External(ref mut node) => node.remove_at(pos),
-            &mut NodeData::Internal(ref mut node) => node.remove_at(pos),
-            &mut NodeData::None => unreachable!(),
-        }
-    }
-    fn is_ext(&self) -> bool {
-        match self {
-            &NodeData::External(_) => true,
-            &NodeData::Internal(_) => false,
-            &NodeData::None => panic!(),
-        }
-    }
-    fn first_key(&self) -> EntryKey {
-        if self.is_ext() {
-            self.extnode().keys[0].to_owned()
-        } else {
-            self.innode().keys[0].to_owned()
-        }
-    }
-    fn len(&self) -> usize {
-        if self.is_ext() {
-            self.extnode().len
-        } else {
-            self.innode().len
-        }
-    }
-    fn is_half_full(&self) -> bool {
-        self.len() >= NUM_KEYS / 2
-    }
-    fn cannot_merge(&self) -> bool {
-        self.len() > NUM_KEYS / 2
-    }
-    fn extnode_mut(&mut self) -> &mut ExtNode {
-        match self {
-            &mut NodeData::External(ref mut node) => node,
-            _ => unreachable!(self.type_name()),
-        }
-    }
-    fn innode_mut(&mut self) -> &mut InNode {
-        match self {
-            &mut NodeData::Internal(ref mut n) => n,
-            _ => unreachable!(self.type_name()),
-        }
-    }
-    pub fn extnode(&self) -> &ExtNode {
-        match self {
-            &NodeData::External(ref node) => node,
-            _ => unreachable!(self.type_name()),
-        }
-    }
-    pub fn ext_id(&self) -> Id {
-        match self {
-            &NodeData::External(ref node) => node.id,
-            &NodeData::None => Id::unit_id(),
-            &NodeData::Internal(ref node) =>  unreachable!(self.type_name()),
-        }
-    }
-    pub fn innode(&self) -> &InNode {
-        match self {
-            &NodeData::Internal(ref n) => n,
-            _ => unreachable!(self.type_name()),
-        }
-    }
-    pub fn type_name(&self) -> &'static str {
-        match self {
-            &NodeData::Internal(_) => "internal",
-            &NodeData::External(_) => "external",
-            &NodeData::None => "none",
-        }
-    }
-    pub fn key_at_right_node(&self, key: &EntryKey) -> Option<&NodeCellRef> {
-        return None;
-        if self.len() > 0 {
-            match self {
-                &NodeData::Internal(ref n) => {
-                    if &n.keys[n.len - 1] < key {
-                        let right_node = n.right.read_unchecked();
-                        if !right_node.is_none() {
-                            let right_innode = right_node.innode();
-                            if right_innode.keys.len() > 0 && &right_innode.keys[0] < key {
-                                debug!("found key to put to right internal page {:?}/{:?}", key, &right_innode.keys[0]);
-                                return Some(&n.right);
-                            }
-                        }
-                    }
-                }
-                &NodeData::External(ref n) => {
-                    if &n.keys[n.len - 1] < key {
-                        let right_node = n.next.read_unchecked();
-                        if !right_node.is_none() {
-                            let right_extnode = right_node.extnode();
-                            if right_extnode.keys.len() > 0 && &right_extnode.keys[0] < key {
-                                debug!("found key to put to right external page {:?}/{:?}", key, &right_extnode.keys[0]);
-                                return Some(&n.next);
-                            }
-                        }
-                    }
-                }
-                _ => unreachable!(),
-            }
-        }
-        return None;
-    }
-}
-
-pub fn write_key_page(search_page: NodeWriteGuard, key: &EntryKey) -> NodeWriteGuard {
-    return search_page;
-    if search_page.len() > 0 {
-        match &*search_page {
-            &NodeData::Internal(ref n) => {
-                if &n.keys[n.len - 1] < key {
-                    let right_ref = &n.right;
-                    debug!("Obtain latch for internal write key page");
-                    let right_node = right_ref.write();
-                    if !right_node.is_none()
-                        && right_node.len() > 0
-                        && &right_node.innode().keys[0] < key
-                    {
-                        debug_assert!(!right_node.is_ext());
-                        debug!("will write to right internal page");
-                        return write_key_page(right_node, key);
-                    }
-                }
-            }
-            &NodeData::External(ref n) => {
-                if &n.keys[n.len - 1] < key {
-                    let right_ref = &n.next;
-                    debug!("Obtain latch for external write key page");
-                    let right_node = right_ref.write();
-                    if !right_node.is_none()
-                        && right_node.len() > 0
-                        && &right_node.extnode().keys[0] < key
-                    {
-                        debug_assert!(right_node.is_ext());
-                        debug!("will write to right external page");
-                        return write_key_page(right_node, key);
-                    }
-                }
-            }
-            _ => unreachable!(),
-        }
-        return search_page;
-    } else {
-        return search_page;
-    }
-}
-
-const LATCH_FLAG: usize = !(!0 >> 1);
-
-pub struct Node {
-    data: UnsafeCell<NodeData>,
-    cc: AtomicUsize,
-}
-
-impl Node {
-    fn new(data: NodeData) -> Self {
-        Node {
-            data: UnsafeCell::new(data),
-            cc: AtomicUsize::new(0),
-        }
-    }
-
-    fn internal(innode: InNode) -> Self {
-        Self::new(NodeData::Internal(box innode))
-    }
-
-    fn external(extnode: ExtNode) -> Self {
-        Self::new(NodeData::External(box extnode))
-    }
-
-    fn none() -> Self {
-        Self::new(NodeData::None)
-    }
-    pub fn none_ref() -> NodeCellRef {
-        Arc::new(Node::none())
-    }
-    pub fn new_external(id: Id) -> Self {
-        Self::external(ExtNode::new(id))
-    }
-
-    pub fn write(&self) -> NodeWriteGuard {
-        let cc = &self.cc;
-        loop {
-            let cc_num = cc.load(Relaxed);
-            let expected = cc_num & (!LATCH_FLAG);
-            if cc.compare_and_swap(expected, cc_num | LATCH_FLAG, Relaxed) == expected {
-                return NodeWriteGuard {
-                    data: self.data.get(),
-                    cc: &self.cc as *const AtomicUsize,
-                    version: node_version(cc_num),
-                };
-            }
-            debug!("acquire latch failed, retry {:b}", cc_num);
-        }
-    }
-
-    pub fn read<'a, F: FnMut(&NodeReadHandler) -> R + 'a, R: 'a>(&'a self, mut func: F) -> R {
-        let mut handler = NodeReadHandler {
-            ptr: self.data.get(),
-            version: 0,
-        };
-        let cc = &self.cc;
-        loop {
-            let cc_num = cc.load(Relaxed);
-            if cc_num & LATCH_FLAG == LATCH_FLAG {
-                debug!("read have a latch, retry {:b}", cc_num);
-                continue;
-            }
-            handler.version = cc_num & (!LATCH_FLAG);
-            let res = func(&handler);
-            let new_cc_num = cc.load(Relaxed);
-            if new_cc_num == cc_num {
-                return res;
-            }
-            debug!("read version changed, retry {:b}", cc_num);
-        }
-    }
-
-    pub fn read_unchecked(&self) -> &NodeData {
-        unsafe { &*self.data.get() }
-    }
-
-    pub fn version(&self) -> usize {
-        node_version(self.cc.load(Relaxed))
-    }
-}
-
-pub fn node_version(cc_num: usize) -> usize {
-    cc_num & (!LATCH_FLAG)
-}
-
-pub struct NodeWriteGuard {
-    data: *mut NodeData,
-    cc: *const AtomicUsize,
-    version: usize,
-}
-
-impl Deref for NodeWriteGuard {
-    type Target = NodeData;
-
-    fn deref(&self) -> &<Self as Deref>::Target {
-        debug_assert_ne!(self.data as usize, 0);
-        unsafe { &*self.data }
-    }
-}
-
-impl DerefMut for NodeWriteGuard {
-    fn deref_mut(&mut self) -> &mut <Self as Deref>::Target {
-        debug_assert_ne!(self.data as usize, 0);
-        unsafe { &mut *self.data }
-    }
-}
-
-impl Drop for NodeWriteGuard {
-    fn drop(&mut self) {
-        // cope with null pointer
-        if self.cc as usize != 0 {
-            let cc = unsafe { &*self.cc };
-            let cc_num = cc.load(Relaxed);
-            cc.store((cc_num & (!LATCH_FLAG)) + 1, Relaxed);
-        }
-    }
-}
-
-impl Default for NodeWriteGuard {
-    fn default() -> Self {
-        NodeWriteGuard {
-            data: 0 as *mut NodeData,
-            cc: 0 as *const AtomicUsize,
-            version: 0
-        }
-    }
-}
-
-pub struct NodeReadHandler {
-    ptr: *const NodeData,
-    version: usize,
-}
-
-impl Deref for NodeReadHandler {
-    type Target = NodeData;
-
-    fn deref(&self) -> &<Self as Deref>::Target {
-        unsafe { &*self.ptr }
-    }
-}
-
-impl RTCursor {
-    fn new(pos: usize, page: &NodeCellRef, ordering: Ordering) -> RTCursor {
-        let mut cursor = RTCursor {
-            index: pos,
-            ordering,
-            page: Some(page.clone()),
-        };
-        match ordering {
-            Ordering::Forward => {
-                let len = page.read(|node| node.len());
-                if pos >= len {
-                    cursor.next();
-                }
-            }
-            Ordering::Backward => {}
-        }
-        cursor
-    }
-    fn boxed(self) -> Box<IndexCursor> {
-        box self
-    }
-}
-
-impl IndexCursor for RTCursor {
-    fn next(&mut self) -> bool {
-        if self.page.is_some() {
-            let current_page = self.page.clone().unwrap();
-            current_page.read(|page| {
-                let ext_page = page.extnode();
-                // debug!("Next id with index: {}, length: {}", self.index + 1, ext_page.len);
-                match self.ordering {
-                    Ordering::Forward => {
-                        if self.index + 1 >= page.len() {
-                            // next page
-                            return ext_page.next.read(|next_page| {
-                                if !next_page.is_none() {
-                                    // debug!("Shifting page forward, next page len: {}", next_page.len());
-                                    self.index = 0;
-                                    self.page = Some(ext_page.next.clone());
-                                    return true;
-                                } else {
-                                    // debug!("iterated to end");
-                                    self.page = None;
-                                    return false;
-                                }
-                            });
-                        } else {
-                            self.index += 1;
-                            // debug!("Advancing cursor to index {}", self.index);
-                        }
-                    }
-                    Ordering::Backward => {
-                        if self.index == 0 {
-                            // next page
-                            return ext_page.prev.read(|prev_page| {
-                                if !prev_page.is_none() {
-                                    debug!("Shifting page backward");
-                                    self.page = Some(ext_page.prev.clone());
-                                    self.index = prev_page.len() - 1;
-                                    return true;
-                                } else {
-                                    debug!("iterated to end");
-                                    self.page = None;
-                                    return false;
-                                }
-                            });
-                        } else {
-                            self.index -= 1;
-                            // debug!("Advancing cursor to index {}", self.index);
-                        }
-                    }
-                }
-                true
-            })
-        } else {
-            false
-        }
-    }
-
-    fn current(&self) -> Option<&EntryKey> {
-        if let &Some(ref page_ref) = &self.page {
-            page_ref.read(|handler| {
-                let page = unsafe { &*handler.ptr };
-                if self.index >= page.len() {
-                    panic!("cursor position overflowed {}/{}", self.index, page.len())
-                }
-                Some(&page.extnode().keys[self.index])
-            })
-        } else {
-            return None;
-        }
-    }
-}
-
-fn insert_into_split<T, S>(
-    item: T,
-    x: &mut S,
-    y: &mut S,
-    xlen: &mut usize,
-    ylen: &mut usize,
-    pos: usize,
-) where
-    S: Slice<Item = T> + BTreeSlice<T>,
-    T: Default,
-{
-    debug!(
-        "insert into split left len {}, right len {}, pos {}",
-        xlen, ylen, pos
-    );
-    if pos < *xlen {
-        debug!("insert into left part, pos: {}", pos);
-        x.insert_at(item, pos, xlen);
-    } else {
-        let right_pos = pos - *xlen;
-        debug!("insert into right part, pos: {}", right_pos);
-        y.insert_at(item, right_pos, ylen);
-    }
-}
-
-pub struct BPlusTreeMergingPage {
-    page: RwLockReadGuard<ExtNode>,
-    mapper: Arc<ExtNodeCacheMap>,
-    pages: Rc<RefCell<Vec<Id>>>,
-}
-
-impl MergingPage for BPlusTreeMergingPage {
-    fn next(&self) -> Box<MergingPage> {
-        unimplemented!()
-    }
-
-    fn keys(&self) -> &[EntryKey] {
-        &self.page.keys[..self.page.len]
-    }
-}
-
-impl MergeableTree for BPlusTree {
-    fn prepare_level_merge(&self) -> Box<MergingTreeGuard> {
-        unimplemented!()
-    }
-
-    fn elements(&self) -> usize {
-        self.len()
-    }
-}
-
-pub struct BPlusTreeMergingTreeGuard {
-    cache: Arc<ExtNodeCacheMap>,
-    txn: RefCell<Txn>,
-    last_node: Arc<Node>,
-    last_node_ref: TxnValRef,
-    storage: Arc<AsyncClient>,
-    pages: Rc<RefCell<Vec<Id>>>,
-}
-
-impl MergingTreeGuard for BPlusTreeMergingTreeGuard {
-    fn remove_pages(&self, pages: &[&[EntryKey]]) {
-        unimplemented!()
-    }
-
-    fn last_page(&self) -> Box<MergingPage> {
-        unimplemented!()
-    }
-}
-
-impl Drop for BPlusTreeMergingTreeGuard {
-    fn drop(&mut self) {
-        self.txn.borrow_mut().commit();
-    }
-}
-
 macro_rules! impl_btree_slice {
     ($t: ty, $et: ty, $n: expr) => {
         impl_slice_ops!($t, $et, $n);
@@ -999,18 +472,7 @@ macro_rules! impl_btree_slice {
 impl_btree_slice!(EntryKeySlice, EntryKey, NUM_KEYS);
 impl_btree_slice!(NodePtrCellSlice, NodeCellRef, NUM_PTRS);
 
-struct RemoveStatus {
-    item_found: bool,
-    removed: bool,
-}
-
-impl Default for Node {
-    fn default() -> Self {
-        Node::none()
-    }
-}
-
-trait BTreeSlice<T>: Sized + Slice<Item = T>
+pub trait BTreeSlice<T>: Sized + Slice<Item = T>
 where
     T: Default,
 {
@@ -1057,7 +519,7 @@ where
 
 #[cfg(test)]
 pub mod test {
-    use super::{BPlusTree, Node, Ordering};
+    use super::*;
     use bifrost::utils::fut_exec::wait;
     use byteorder::BigEndian;
     use byteorder::WriteBytesExt;
@@ -1088,6 +550,7 @@ pub mod test {
     use std::time::Duration;
     use itertools::Itertools;
     use index::EntryKey;
+    use index::btree::node::*;
 
     extern crate env_logger;
     extern crate serde_json;
